// src/app/models/baseline.model.ts
export class BaselineOutput {
    cosineSim: number = 0.7919469343423844;
    jaccardWords: number = 0.035230973085365005;
    jaccardBigrams: number = 0.007854583366040075;
    tfIdfSim: number = 0.19162804064123354;
    normEditDist: number = 0.03616429257168752;
    lenDif: number = 2736.466;
    lenRatio: number = 0.03778615810327395;
}